+++ CHANGELOG.md
@@ -5,8 +5,10 @@
 - Constrain noise source more accurately for `generateCylinder`.
 
 ### Fixed
-- Add size scaling to `generateLine`.
 - Use correct operator for 3D noise `y0` calculation.
+
+### Removed
+- Remove white noise scaling using `frequency` (makes no difference).
 
 ## [v0.1] – 2017-02-14
 ### Added